+++ test/sparql/query/result/query_result_test.exs
@@ -21,11 +21,18 @@ defmodule SPARQL.Query.ResultTest do
       }
     ]}
 
-  test "get/2" do
-    assert Query.Result.get(@example, "p") ==
-             [~I<http://example.org/p1>, ~I<http://example.org/p2>]
-    assert Query.Result.get(@example, :p) ==
-             [~I<http://example.org/p1>, ~I<http://example.org/p2>]
+  describe "get/2" do
+    test "when values for the given variable exists" do
+      assert Query.Result.get(@example, "p") ==
+               [~I<http://example.org/p1>, ~I<http://example.org/p2>]
+      assert Query.Result.get(@example, :p) ==
+               [~I<http://example.org/p1>, ~I<http://example.org/p2>]
+    end
+
+    test "when no values for the given variable exist" do
+      assert Query.Result.get(@example, "foo") == nil
+      assert Query.Result.get(@example, :foo) == nil
+    end
   end
 
 end